+++ tools/generate-fhir/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - Properly generate profile urls automatically
+- Fix value mappings for top level codeable concepts
 
 ## 0.7.2